Port the gui, container, loot and time apis, 231 to 200
Effect icons draw by sprite id through a render pipeline now, with the alpha packed into a colour
rather than set as gui state, which is what removed setColor. ClickType became ContainerInput and
Container's open hooks take a ContainerUser. CustomData#getUnsafe became copyTag, and ItemStack
save and parse go through its codecs.

The moon phase became a positional environment attribute rather than a level wide number, and
Level#getDayTime became the dimension's own clock.

+++ src/main/java/se/mickelus/tetra/effect/data/provider/number/EntityPropertyNumberProvider.java
@@ -27,7 +27,7 @@ public class EntityPropertyNumberProvider implements NumberProvider {
             case burningDuration -> resolvedEntity.getRemainingFireTicks();
             case airSupply -> resolvedEntity.getAirSupply();
             case maxAirSupply -> resolvedEntity.getMaxAirSupply();
-            case fallDistance -> resolvedEntity.fallDistance;
+            case fallDistance -> (float) resolvedEntity.fallDistance;
             case health -> resolvedEntity instanceof LivingEntity livingEntity ? livingEntity.getHealth() : 0;
             case maxHealth -> resolvedEntity instanceof LivingEntity livingEntity ? livingEntity.getMaxHealth() : 0;
             case armor -> resolvedEntity instanceof LivingEntity livingEntity ? livingEntity.getArmorValue() : 0;
